Export the embedding model with a bfloat16 feature boundary
Both sides of embedding.onnx already round to bfloat16: the vision model
emits bf16 inside its DD nodes and widens the features back only to hand
them over, and the decoder's first op on each of the four full-length
outputs is a cast to bf16. Exporting the boundary as bf16 removes both
halves of that round trip, which is exact -- bf16 -> fp32 is a widening
and fp32 -> bf16 its inverse, so every value is unchanged.

  - the four feature dummies drive the exported input dtype, so they
    become bf16
  - inputs_embeds is cast right after the embedding lookup, so the
    scatter and all four outputs are bf16
  - the .to(device, dtype) calls become guarded: both arguments are
    static during export and an unconditional .to() traces to a Cast
    even when it is a no-op, which would leave four full-width copies

Measured on Qwen3-VL-4B / RyzenAI, 8k prompt, single image, 13 turns:
host peak 7.706 G -> 7.583 G with the matching partition-side change,
ttft and tps unchanged, generated text byte-identical.

+++ Qwen-Qwen3-VL-4B-Instruct/builtin/user_script.py
@@ -119,16 +119,18 @@ def get_embedding_dummy_inputs(model=None):
         "image_features": torch.randn(
             num_logical_patches,
             out_hidden_size,
-            dtype=torch.float32,  # fp32 to match embedding model export dtype
+            # bf16 in, bf16 out: the vision model already rounds these to bf16 on the NPU and
+            # the scatter runs in bf16, so exporting them as fp32 only buys a cast back down
+            dtype=torch.bfloat16,
         ),
         "deepstack_features_0": torch.randn(
-            num_logical_patches, out_hidden_size, dtype=torch.float32
+            num_logical_patches, out_hidden_size, dtype=torch.bfloat16
         ),
         "deepstack_features_1": torch.randn(
-            num_logical_patches, out_hidden_size, dtype=torch.float32
+            num_logical_patches, out_hidden_size, dtype=torch.bfloat16
         ),
         "deepstack_features_2": torch.randn(
-            num_logical_patches, out_hidden_size, dtype=torch.float32
+            num_logical_patches, out_hidden_size, dtype=torch.bfloat16
         ),
     }
